Fix: 권한에 따른 접근 페이지 설정

+++ src/utils/axios.tsx
@@ -41,30 +41,14 @@ const ApiClient = () => {
     async (err) => {
       const { response: result, config: originalRequest } = err;
       let errMessage: string = result.data.message;
-
-      const adminAuth = localStorage.getItem(LOGIN_AUTH);
-      if (errMessage === 'Unauthenticated' && adminAuth === 'admin') {
-        const retryLogin = async () => {
-          try {
-            // 토큰 교체
-            const _result = await admin.admin_refresh_token();
-            localStorage.setItem(ACCESS_TOKEN, _result.data.accessToken);
-            // 기존에 쏘려고했던 api 재호출
-            originalRequest.headers.Authorization = `Bearer ${_result.data.accessToken}`;
-            return instance(originalRequest);
-          } catch (err) {
-            console.log('err', err);
-            const _page = localStorage.getItem(LOGIN_AUTH);
-            localStorage.clear();
-
-            if (_page === 'user') {
-              window.location.href = ENDPOINT.USER_MAIN;
-            } else {
-              window.location.href = ENDPOINT.ADMIN_DASHBOARD;
-            }
-          }
-        };
-        retryLogin();
+      if (result.status === 401) {
+        const adminAuth = localStorage.getItem(LOGIN_AUTH);
+        localStorage.clear();
+        if (adminAuth === 'admin') {
+          window.location.href = ENDPOINT.ADMIN_LOGIN;
+        } else if (adminAuth === 'user') {
+          window.location.href = ENDPOINT.USER_LOGIN;
+        }
       }
 
       // error response가 떨어지지 않는 경우
